Simplifies sensor logic for non-triangle robots.
Simplifies the `presenceSolGauche` and `presenceSolDroit` methods.

For robots other than PAMI_TRIANGLE, the sensor logic is simplified
to always return `true`, as these sensors are not connected on
these robot types.

Also, updates the shell command output to only log sol gauche/droit
for the triangle robot. This avoids confusion in the shell command
logs for the other robots.