Adds sensor origin to lidar telemeter
Adds the concept of sensor origin to the lidar telemeter interface and implementations.

This allows for the correct calculation of point coordinates relative
to the robot's coordinate system, taking into account the physical
placement of the lidar sensor on the robot.

The sensor origin can be set and retrieved, and is used to offset
the calculated point coordinates in the LidarService. This ensures
that the lidar data is correctly aligned with the robot's position
and orientation.