Enables setting Lidar sensor origin
Allows to configure the Lidar sensor origin with x and y coordinates.

This change introduces the ability to specify the Lidar sensor's
origin, enabling more accurate point calculations by accounting
for the sensor's physical placement on the robot. It modifies the
ILidarTelemeter interface and implementing classes to use x and y
coordinates instead of a Point object for setting the origin.
It also updates the LidarService to consider sensor origin when transforming lidar scan data.