Implements product placement calibration for triangle robots
Implements the `calagePriseProduitSolAvant` method for triangle robots.

It checks if the minimum number of mandatory ground sensors are detecting the ground before attempting to pick up a product.
An exception is thrown if the number of mandatory sensors exceeds 2, as the robot is configured with a maximum of two ground sensors.