Improves PAMI robot functionality
Removes unused code and implements exception for unimplemented methods.

This commit removes commented-out code related to screen functionality in `PamiCommonContext.java`.
It also introduces a `NotImplementedException` in `PamiIOServiceRobot.java` for the `calagePriseProduitSolAvant` method, indicating that the method is not yet implemented for the PAMI robot.
Finally, it adds a missing newline in `LD19LidarTelemeterOverSocket.java`.