Adds super open position for robot fingers
Adds a "super open" position for the robot's fingers to improve object manipulation, specifically when picking up objects from the top position on the first floor.

This new position is used in the front and rear finger servos to allow a wider grip.  Also, the configuration shell commands now use super open by default when setting the finger positions.